Move gameplay-related settings to its own cpp file; Reorganize .ini to fit new "gameplay" section
