+++ dllmain/dllmain.vcxproj
@@ -678,6 +678,7 @@ echo | set /p dummyName=#define GIT_BRANCH ^"%VAR%^" &gt;&gt; gitparams.h</Comma
     <ClCompile Include="EndSceneHook.cpp" />
     <ClCompile Include="ExceptionHandler.cpp" />
     <ClCompile Include="GameFlags.cpp" />
+    <ClCompile Include="Gameplay.cpp" />
     <ClCompile Include="HDProject.cpp" />
     <ClCompile Include="input.cpp" />
     <ClCompile Include="Game.cpp" />
