+++ dllmain/dllmain.vcxproj.filters
@@ -568,6 +568,9 @@
     <ClCompile Include="TitleMenu.cpp">
       <Filter>dllmain</Filter>
     </ClCompile>
+    <ClCompile Include="Gameplay.cpp">
+      <Filter>dllmain</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\external\inireader\ini_parser.hpp">
@@ -1613,7 +1616,7 @@
     </ClInclude>
     <ClInclude Include="..\external\ImGuiFileDialog\ImGuiFileDialog.h">
       <Filter>external\ImGuiFileDialog</Filter>
-	</ClInclude>
+    </ClInclude>
     <ClInclude Include="..\dxvk\src\dxvk-cache.hpp">
       <Filter>dxvk</Filter>
     </ClInclude>
