+++ settings/settings.ini
@@ -188,6 +188,7 @@ ReloadWithoutZoom = false
 [CONTROLLER]
 ; Change the controller sensitivity. For some reason the vanilla game doesn't have an option to change it for controllers, only for the mouse.
 ; Min 0.5, max 4.0.
+OverrideControllerSensitivity = false
 ControllerSensitivity = 1.0
 
 ; Removes unnecessary deadzones that were added for Xinput controllers.
@@ -196,6 +197,7 @@ RemoveExtraXinputDeadzone = true
 ; Change the Xinput controller deadzone. (Unrelated to the previous option)
 ; The game's default is 1, but that seems unnecessarily large, so we default to 0.4 instead.
 ; Min 0.0, max 3.5.
+OverrideXinputDeadzone = true
 XinputDeadzone = 0.4
 
 ; Removes the need to be aiming the weapon before you can reload it.
@@ -249,6 +251,49 @@ MultithreadFix = true
 ; (not fully tested, could cause issues if level has many models to load!)
 PrecacheModels = false
 
+[GAMEPLAY]
+; Unlocks the JP-only classic camera angles during Ashley's segment.
+AshleyJPCameraAngles = false
+
+; Silence Ashley's armored outfit ("Special 2").
+; For those who also hate the constant "Clank Clank Clank".
+SilenceArmoredAshley = false
+
+; Choose between Normal and Professional difficulty modes when starting a new game of Separate Ways.
+SeparateWaysDifficultyMenu = true
+
+; Unlocks minor difficulty boosts previously exclusive to the NTSC console versions of RE4.
+; Higher starting adaptive difficulty, more difficult Ada missions, and a more difficult Mercenaries village stage.
+; Bottle caps require 3000 points in the shooting gallery, and Easy difficulty is removed from the title menu.
+EnableNTSCMode = false
+
+; Allows Ashley to Suplex enemies in very specific situations.
+; (previously was only possible in the initial NTSC GameCube ver., was patched out in all later ports.)
+AllowAshleySuplex = false
+
+; Allows quickturning character to camera direction when wielding Matilda
+; (only effective if MouseTurning is disabled)
+AllowMatildaQuickturn = true
+
+; Fixes the Ditman glitch, which would allow players to increase their walk speed.
+; Not recommended, but mod makers may be interested in enabling this (see Bin32/re4_tweaks/setting_overrides/overrides_info.txt)
+FixDitmanGlitch = false
+
+; Changes sprint key to act like a toggle instead of needing to be held.
+UseSprintToggle = false
+
+; Adds a screen shake effect when firing a rifle.
+RifleScreenShake = false
+
+; Disables most of the QTEs, making them pass automatically.
+DisableQTE = false
+
+; Unlike the previous option, this only automates the "mashing" QTEs, making them pass automatically. Prompts are still shown!
+AutomaticMashingQTE = false
+
+; Limit the Matilda to one three round burst per trigger pull.
+LimitMatildaBurst = false
+
 [MISC]
 ; Path to DLL to wrap, comment or leave empty to wrap system DLL
 ; (only set this if you need re4_tweaks to 'chain-load' another DLL that makes use of the same filename)
@@ -256,14 +301,6 @@ PrecacheModels = false
 
 ; Disable online updates and update checking.
 NeverCheckForUpdates = false
-
-; Unlocks minor difficulty boosts previously exclusive to the NTSC console versions of RE4.
-; Higher starting adaptive difficulty, more difficult Ada missions, and a more difficult Mercenaries village stage.
-; Bottle caps require 3000 points in the shooting gallery, and Easy difficulty is removed from the title menu.
-EnableNTSCMode = false
-
-; Unlocks the JP-only classic camera angles during Ashley's segment.
-AshleyJPCameraAngles = false
 
 ; Restores the playback of demo videos that play when the game is idle at the main menu or the "press any key" screen for about 20 seconds.
 ; By default, the PC port only includes a single demo video (demo0.sfd/demo0eng.sfd).
@@ -282,9 +319,6 @@ RestoreAnalogTitleScroll = true
 ; Use -1 to leave as your game EXEs default, 0 for low-violence mode (as used in JP/GER version), or 2 for full violence.
 ViolenceLevelOverride = -1
 
-; Adds a screen shake effect when firing a rifle.
-RifleScreenShake = false
-
 ; Allows selling the (normally unused) handgun silencer to the merchant.
 ; If you use a mod that changes the merchant price tables you might want to disable this
 ; But if you're only using RE4HD you can leave this enabled
@@ -293,31 +327,6 @@ AllowSellingHandgunSilencer = true
 ; Allows the game to display Leon's mafia outfit ("Special 2") on cutscenes.
 AllowMafiaLeonCutscenes = true
 
-; Silence Ashley's armored outfit ("Special 2").
-; For those who also hate the constant "Clank Clank Clank".
-SilenceArmoredAshley = false
-
-; Allows Ashley to Suplex enemies in very specific situations.
-; (previously was only possible in the initial NTSC GameCube ver., was patched out in all later ports.)
-AllowAshleySuplex = false
-
-; Allows quickturning character to camera direction when wielding Matilda
-; (only effective if MouseTurning is disabled)
-AllowMatildaQuickturn = true
-
-; Fixes the Ditman glitch, which would allow players to increase their walk speed.
-; Not recommended, but mod makers may be interested in enabling this (see Bin32/re4_tweaks/setting_overrides/overrides_info.txt)
-FixDitmanGlitch = false
-
-; Changes sprint key to act like a toggle instead of needing to be held.
-UseSprintToggle = false
-
-; Disables most of the QTEs, making them pass automatically.
-DisableQTE = false
-
-; Unlike the previous option, this only automates the "mashing" QTEs, making them pass automatically. Prompts are still shown!
-AutomaticMashingQTE = false
-
 ; Whether to skip the Capcom etc intro logos when starting the game.
 SkipIntroLogos = false
 
@@ -325,12 +334,6 @@ SkipIntroLogos = false
 ; Mod authors might prefer intro logos to be skipped without skipping the menu logo, this can be used in a mod override-ini if desired.
 ; (will only take effect when SkipIntroLogos is also set to true)
 SkipMenuFades = false
-
-; Limit the Matilda to one three round burst per trigger pull.
-LimitMatildaBurst = false
-
-; Choose between Normal and Professional difficulty modes when starting a new game of Separate Ways.
-SeparateWaysDifficultyMenu = true
 
 ; After beating the game, force the game to always show the original main menu background image of Leon and Ashley.
 AlwaysShowOriginalTitleBackground = false
